[not verified] publicize: unify from component
