+++ projects/plugins/jetpack/extensions/plugins/publicize/components/form/index.js
@@ -1,68 +1,62 @@
 /**
- * Higher Order Publicize sharing form composition.
+ * Publicize sharing form component.
  *
- * Uses Gutenberg data API to dispatch publicize form data to
- * editor post data in format to match 'publicize' field schema.
+ * Displays text area and connection list to allow user
+ * to select connections to share to and write a custom
+ * sharing message.
  */
-
-/**
- * External dependencies
- */
-import { compose } from '@wordpress/compose';
-import { withSelect, withDispatch } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
-import PublicizeFormUnwrapped from '../form-unwrapped';
+import PublicizeConnection from '../connection';
+import PublicizeSettingsButton from '../settings-button';
+import MessageBox from '../message-box';
+import useSocialMediaConnections from '../../hooks/use-social-media-connections';
+import useSelectSocialMedia from '../../hooks/use-select-social-media';
 
-const PublicizeForm = compose( [
-	withSelect( select => {
-		return {
-			connections: select( 'core/editor' ).getEditedPostAttribute(
-				'jetpack_publicize_connections'
-			),
-			shareMessage: select( 'jetpack/publicize' ).getShareMessage(),
-			maxLength: select( 'jetpack/publicize' ).getShareMessageMaxLength(),
-		};
-	} ),
-	withDispatch( ( dispatch, { connections } ) => ( {
-		/**
-		 * Toggle connection enable/disable state based on checkbox.
-		 *
-		 * Saves enable/disable value to connections property in editor
-		 * in field 'jetpack_publicize_connections'.
-		 *
-		 * @param {number}  id - ID of the connection being enabled/disabled
-		 */
-		toggleConnection( id ) {
-			const newConnections = connections.map( connection => ( {
-				...connection,
-				enabled: connection.id === id ? ! connection.enabled : connection.enabled,
-			} ) );
+export default function PublicizeForm( { refreshCallback } ) {
+	const { connections, message, maxLength } = useSelectSocialMedia();
+	const { toggleEnableState, updateMessage } = useSocialMediaConnections();
 
-			dispatch( 'core/editor' ).editPost( {
-				jetpack_publicize_connections: newConnections,
-			} );
-		},
+	/**
+	 * Check to see if form should be disabled.
+	 *
+	 * Checks full connection list to determine if all are disabled.
+	 * If they all are, it returns true to disable whole form.
+	 *
+	 * @returns {boolean} True if whole form should be disabled.
+	 */
+	function isDisabled() {
+		return connections.every( connection => ! connection.toggleable );
+	}
 
-		/**
-		 * Handler for when sharing message is edited.
-		 *
-		 * Saves edited message to state and to the editor
-		 * in field 'jetpack_publicize_message'.
-		 *
-		 * @param {object} message - the sharing message.
-		 */
-		onMessageChange( message ) {
-			dispatch( 'core/editor' ).editPost( {
-				meta: {
-					jetpack_publicize_message: message,
-					jetpack_publicize_hasEditedShareMessage: true,
-				},
-			} );
-		},
-	} ) ),
-] )( PublicizeFormUnwrapped );
+	return (
+		<div id="publicize-form">
+			<ul className="jetpack-publicize__connections-list">
+				{ connections.map( ( { display_name, enabled, id, service_name, toggleable } ) => (
+					<PublicizeConnection
+						disabled={ ! toggleable }
+						enabled={ enabled }
+						key={ id }
+						id={ id }
+						label={ display_name }
+						name={ service_name }
+						toggleConnection={ toggleEnableState }
+					/>
+				) ) }
+			</ul>
 
-export default PublicizeForm;
+			<PublicizeSettingsButton refreshCallback={ refreshCallback } />
+
+			{ connections.some( connection => connection.enabled ) && (
+				<MessageBox
+					disabled={ isDisabled() }
+					maxLength={ maxLength }
+					onChange={ updateMessage }
+					message={ message }
+				/>
+			) }
+		</div>
+	);
+}
